+++ src/openassetio-core/hostAPI/HostInterface.cpp
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2013-2022 The Foundry Visionmongers Ltd
+#include <openassetio/hostAPI/HostInterface.hpp>
+#include <openassetio/typedefs.hpp>
+
+namespace openassetio {
+inline namespace OPENASSETIO_CORE_ABI_VERSION {
+namespace hostAPI {
+InfoDictionary HostInterface::info() const { return {}; }
+}  // namespace hostAPI
+}  // namespace OPENASSETIO_CORE_ABI_VERSION
+}  // namespace openassetio